requirements.txt
requests
Flask
python-dotenv
redis 
apscheduler
dash